Extract base dialog
